Fix local build infrastructure files
We had several problems when running `docker-compose build`, which build
the containers, and consequently `docker-compose up`. First off,
`docker-compose up` flat out didn't work in the way the repo was set up
previously and we couldn't build the local container either.

We also updated the Ruby image in the Dockerfile to reflect the version
specified in the Gemfile. Besides that, we included a .dockerignore
file, which speeds up the build process.

This whole fix involved a series of steps. We adjusted Bundler, the OS
the Ruby image uses and Yarn installations, and the database.yml.example
file. Now we should be able to bring up a container from zero in our
local environment without any problems.